requirements.txt
Flask==3.0.0
requests
google-cloud-datastore==2.15.1
authlib 
dotenv
circle-developer-controlled-wallets